check err_code when unsubscribe
