+++ src/app/integrations/dashamail/exceptions.py
@@ -12,3 +12,7 @@ class DashamailWrongResponse(DashamailHTTPException):
 
 class DashamailSubscriptionFailed(DashamailException):
     pass
+
+
+class DashamailUnsubscriptionFailed(DashamailException):
+    pass
